# Conflicts:
#	shiying-config-respo/respo/shiying-govern-gateway-dev.yml
#	shiying-service-test_projects/src/main/java/cn/shiying/test_projects/mapper/TestProjectsMapper.java

+++ shiying-common/src/main/java/cn/shiying/common/entity/Drugs/DrugsPurchaseDetailed.java
@@ -41,7 +41,7 @@ public class DrugsPurchaseDetailed implements Serializable {
     /**
      * 进货单价
      */
-    private Double pdMoney;
+    private Integer pdMoney;
 
     /**
      * 进货数量
